app/requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0

# Vector store
chromadb>=0.4.0

# Google AI
google-generativeai>=0.3.0

# Utilities
pydantic>=2.0.0
